requirements.txt
flask==3.0.0
google-cloud-logging==3.9.0
gunicorn==20.1.0